@@ -5,7 +5,7 @@ description = ""
 authors = ["Moritz Boos <m.boos@eyeo.com>"]
 
 [tool.poetry.dependencies]
-python = ">3.8,<3.11"
+python = "^3.10"
 numpy = "^1.23.1"
 scipy = "^1.8.1"
 praw = "^7.6.0"
@@ -13,11 +13,21 @@ matplotlib = "^3.5.2"
 seaborn = "^0.11.2"
 pandas = "^1.4.3"
 ipykernel = "^6.15.1"
+typer = "^0.7.0"
+psaw = "^0.1.0"
+beautifulsoup4 = "^4.11.1"
 
 [tool.poetry.dev-dependencies]
 pytest = "^5.2"
 stubs = "^1.0.0"
 joblib = "^1.1.0"
+
+[tool.poetry.group.dev.dependencies]
+mypy = "^0.991"
+black = {version = "^22.12.0", allow-prereleases = true}
+
+[tool.poetry.scripts]
+ul-scraper = "ultralight_hiking.cmd:run"
 
 [build-system]
 requires = ["poetry-core>=1.0.0"]
